+++ AddOns/Syndicator/Tracking/Tooltips.lua
@@ -54,7 +54,7 @@ function Syndicator.Tooltips.AddItemLines(tooltip, summaries, itemLink)
     return
   end
 
-  local connectedRealmsOnly = Syndicator.Config.Get("tooltips_connected_realms_only")
+  local connectedRealmsOnly = Syndicator.Config.Get("tooltips_connected_realms_only_2")
   local factionOnly = Syndicator.Config.Get("tooltips_faction_only")
   if IsBoA(itemLink) then
     connectedRealmsOnly = false
@@ -206,7 +206,7 @@ function Syndicator.Tooltips.AddCurrencyLines(tooltip, currencyID)
     return
   end
 
-  local summary = Syndicator.Tracking.GetCurrencyTooltipData(currencyID, Syndicator.Config.Get("tooltips_connected_realms_only"), Syndicator.Config.Get("tooltips_faction_only"))
+  local summary = Syndicator.Tracking.GetCurrencyTooltipData(currencyID, Syndicator.Config.Get("tooltips_connected_realms_only_2"), Syndicator.Config.Get("tooltips_faction_only"))
 
   if Syndicator.Config.Get("tooltips_sort_by_name") then
     table.sort(summary, CharacterAndRealmComparator)