+++ AddOns/Syndicator/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Syndicator
 
-## [84](https://github.com/Baganator/Syndicator/tree/84) (2024-07-31)
-[Full Changelog](https://github.com/Baganator/Syndicator/compare/83...84) 
+## [102](https://github.com/Baganator/Syndicator/tree/102) (2024-08-14)
+[Full Changelog](https://github.com/Baganator/Syndicator/compare/101...102) 
 
-- Remove 10.2.7 from toc  
+- Rework guild realm detection to account for missing realm data from player names